requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
sqlalchemy[asyncio]>=2.0.0
alembic>=1.11.1
asyncpg>=0.28.0
psycopg2-binary>=2.9.6
pydantic-settings>=2.0.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.1
ruff>=0.0.280
redis>=5.0.0
celery>=5.3.1
python-magic>=0.4.27
python-magic-bin>=0.4.14
tenacity>=8.2.3
google-generativeai>=0.3.1
slowapi>=0.1.8
email-validator>=2.1.0
aiosqlite>=0.19.0
PyPDF2>=3.0.0
pypdf>=3.17.0
python-docx>=1.1.0
pytesseract>=0.3.10
pdf2image>=1.17.0
Pillow>=10.0.0